Move addNewCourse action to own file
